# [Java.Interop] use StringComparer.Ordinal for Dictionary